+++ tests/test_util.py
@@ -1,8 +1,12 @@
+from pathlib import Path
+
 import jmespath
 import jq
 import jsonpointer
+import orjsonl
 import pytest
 import transon
+from tikray.util.data import load_json, save_json
 from tikray.util.expression import compile_expression
 from tikray.util.locator import to_pointer
 
@@ -48,3 +52,17 @@ def test_compile_expression_unknown():
     with pytest.raises(TypeError) as ex:
         compile_expression(type="foobar", expression=None)
     assert ex.match("Compilation failed. Type must be either jmes or jq or transon: foobar")
+
+
+def test_load_jsonl_by_suffix(tmp_path: Path):
+    data = load_json(Path("examples/eai-warehouse.json"))
+    tmp_path = tmp_path / "testdrive.jsonl"
+    orjsonl.save(tmp_path, [data])
+    assert load_json(tmp_path) == [data]
+
+
+def test_load_jsonl_by_flag(tmp_path: Path):
+    data = load_json(Path("examples/eai-warehouse.json"))
+    tmp_path = tmp_path / "testdrive.json"
+    save_json([data], tmp_path, use_jsonl=True)
+    assert load_json(tmp_path, use_jsonl=True) == [data]